New working project code
These 2 weeks we have been slaving away with coding and this week i was the designated coder while the other group members were backseat coding. Improvements that could be made is, symbols and air presure scale on the graph. We had a lot of problems with our original code and decided to redo the whole coding part to make it easier.